flask_crud/requirements.txt
Flask==2.0.1
Flask-MySQLdb==2.0.0
mysqlclient==2.1.0
python-dotenv==0.19.0